fixes small compilation errors and updates README to include information on the python API

+++ tests/test_labelblk.cpp
@@ -69,7 +69,7 @@ int main(int argc, char** argv)
         // ** Write and read labelblk data **
         
         // create label 64 image volume
-        unsigned long long * img_labels = new unsigned long long [BLK_SIZE*BLK_SIZE*BLK_SIZE];
+        uint64 * img_labels = new uint64 [BLK_SIZE*BLK_SIZE*BLK_SIZE];
         for (int i = 0; i < (LWIDTH*LHEIGHT); ++i) {
             img_labels[i] = limg1_mask[i];
             img_labels[i+BLK_SIZE*BLK_SIZE] = limg2_mask[i];
@@ -86,7 +86,7 @@ int main(int argc, char** argv)
         // retrieve the image volume and make sure it makes the posted volume
         Labels3D labelcomp = dvid_node.get_labels3D(label_datatype_name, lsizes, start);
         
-        const unsigned long long* labeldatacomp = labelcomp.get_raw();
+        const uint64* labeldatacomp = labelcomp.get_raw();
         for (int i = 0; i < BLK_SIZE*BLK_SIZE*BLK_SIZE; ++i) {
             if (labeldatacomp[i] != img_labels[i]) {
                 cerr << "Read/write mismatch" << endl;
